Fix a nasty, nasty integer arithmetic bug
When I was shifting 1 right, charm (?) was somehow not converting it to
a long integer as it should, so there was an overflow and the
**supposedly** positive integer 1 << 63 was turning negative! But only
WITHIN charm! I have fixed this by creating something that absolutely
must be a long integer, and adding that to the appropriate shit. Since
it's beyond the bits I care about, it only adds computational time and
nothing more.