Incorporate unFlipIntFixed function to finally fix region integer flip
This function pairs correctly with activeHyperplaneSet to return those
functions from the **POSET-SPECIFIED INEQUALITIES** that DON'T need to
be 'flipped' to be true. By my own unfortunate convention -- with a
double negative -- the poset code defaults to Wx + b <= 0. So those
inequalities that don't have to be flipped to be true with respect to
this default, must then BE FLIPPED in order to be true with respect to
the convention used other places in the code, viz. Wx + b >= 0. This
explains the need for negation on lines 115 and 116 in
NodeCheckerGenericReach.py.

The TLLReach version (in NodeCheckerLowerBdVerify.py) seems to be
working with these combined changes, but I have to double check the
logic there.