Fix bug in processNodeSuccessors related to degenerate nodes
In one example, found in index [1][0] of the "TLL experiment file"
"TLLExperimentGroup20210225192514.p" processNodeSuccessors returned a
degenerate node. That is for a specific set of inequalities, CDD
returned a 'linear' pair after canonicalizing; this means that some of
the provided inequalities were degenerate, and we need to ignore these.

Such nodes can be identified by their .faceInt property set to zero.