Preliminary support for LTI systems
This adds Asys and Bsys parameters to specify LTI system matrices.
When these parameters are both set, the output constraints are
interpreted as constraints on the next state for an LTI system using
the specified TLL network as a controller, i.e.

x[t+1] = A * x[t] + B * TLL(x[t])