Add code to 'manually' find minimal H-rep. when CDD gives nonsensical answer
This code is only triggered when CDD returns 'linear' constraints, i.e.
when it thinks that some constraints create an un-bounded linear region
of lower dimension. This is incorrect in all instances, so we simply
write a manual min H-rep finder to replace CDD's results.

Note 1: it is slower.

Note 2: it doesn't universally solve the problem. It seems CDD is
returning incorrect results, even when it's not finding so-called linear
regions!