Add populatePoset option find flippable faces only
Since we technically only need the flippable (i.e. non-flipped) faces in
the minimum H representation to find the successor nodes, add this
option to stop once all of those faces are found. This saves LP calls in
the minHRep calculation, and reduces the total runtime needed to
populate the poset.

However, it comes at the price of no feeding non-redudant faces to the
node checker, which increases the number of constraints in its LPs.
However, this should be worth the tradeoff since adding constraints
isn't as bad as doing multiple LPs (this will also be offset by the new,
faster LP solver, Clp).