Modify minHRep convex programs to catch active but not binding constraints
The new convex program both pushes the constraint of interest out, away
from the feasible set by adjusting its constant **and** it keeps the original
contraint but **flips** it. Thus, if we get an optimal point that
satisfies all of the original constraints, including the **unflipped** version
of the constraint of interest, we can conclude that optimal point was exactly
at a corner or something -- i.e. the constraint we pushed out was active but
not binding.