Add slack to fix numerical errors (see bounding box error)
NOTE: we should rewind this new way of checking for redundant hyperplanes
if  it requires this kind of adjustment. It's ok to allow this kind of
slack for the bounding box calculation, since the worst thing that happens
there is that we include a few extra hyperplances. But this rounding
affects the soundness of the algorithm.