Add safe option to over-compute number of valid regions
This option uses a different LP that may return "corners" of valid
regions that are also counted as full regions themselves. Thus, this
option may force a verifier to effectively check some regions twice,
but with benefit that all regions are guaranteed to be checked.

**Note: this only seems to matter when numerial issues come into play
e.g. when some hyperplanes are so close to intersecting that an LP
can't tell whether they do or not.