Apparently working implementation of a distributed hash table
The key hold up in getting this implementation done was the structure
of having multiple channels going to the same hash worker. The problem
is in making sure that all of the messages are successfully read one at
a time, so that there are no collisions on the hash table.

The key to the implementation was to spawn one proxy method for each of
these channels that listens only to its own channel. When it receives
a message, it signals the main listener with its channel inded via a
LOOPBACK channel (!!) and sets a future for the main listener to set
once it has processed that message. The main listener reads from the
loopback channel and processes the messages one at a time, telling the
individual channel listeners to resume once it's done. This ensures
synchrony.

There is a ton of dead code that is commented out. This will be
commented out in a subsequent commit.