Allow splitting of poset and hash workers on PEs
This sort of ridiculous solution is the result of groups created on
specific PEs not working correctly. Specifically, if you create a group
in such a way, you will be unable create channels from elements of that
group because the group is now a "section proxy". This seems like a bug
or at least undocumented behavior. (Surprise, surprise)

The simplest solution given my current code, was to just create groups
on ALL of the PEs, and then use sections to do the scheduling. So
effectively the poset group has idle Chares sitting on the PEs used
by the hash group's main workers and vise versa.

TO DO: add logic to ensure that the specified PE lists don't overlap?

[[NOTE: the PE lists to have to be contiguous, to make life easier]]