Implement new architecture for init/update/check of nodes as created
The new structure has the DistributedHash Chare taking two new arguments: a node class and a Group proxy.

The main idea is that the node class's constructor -- which SHOLD NOT BE OVERRIDDEN -- will be called on the creation of a node as usual; but the supplied class can implement 'init', 'update' and 'check' methods as desired, and if present they will be executed at the correct time.

The Group proxy will be made available inside an instance of the node class by way of a self.localProxy pointing to the correct proxy for the current PE. This means that the Group can be effectively used to distribute common information to all of the PEs and made available to the node instances WITHOUT expensive copies every time a node is created.

This scheme allows the 'unflipping' of node integers to happen concurrently, for example. And it also allows checking to be implemented relatively easily: that is, by sub-classing and implementing a single method.

Moreover, the incorporation of an update method should allow us to handle future requirements for the TLL compiler without much trouble.

Finally, note that the parent hash worker Chare hosting the node is also made available inside the node, via the self.parentChare property. This, combined with the user-supplied proxy should allow fairly sophisticated extensions without having to write much additional code.

All in all, this should be a win for the programmer, while not incurring much performance hit.

(Although as I have already noted, the distributed hash table implementation definitely does incur SOME performance hit over the old version. But here's hoping the ability to terminate in the middle of processing a poset level will more than offset that cost. This is the next implementation task.)