Early termination for sucessor workers/sucessor worker yielding for parallelism
Apparently, pycharm does not have preemptive scheduling, and the only way that a Chare's method can run on a PE is if all other remote methods have yielded (say, by waiting for a Future). Thus, when the successor worker and hash workers were running on the same PEs, the old version was effectively computing all of the successors for the level and *then* hashing them. This would obviously delay the checking until the level was done, which kind of defeats the point. (Of course this doesn't apply if the successor/hash workers are assigned to disjoint PEs).

Since, I had to implement the early termination logic anyway, I decided to use this channel to signal when a successor worker should yield control on of the hash workers.

There is a significant performance hit here, that seems to be coming from the addition of feedback channels only to the DistributedHash Chare, not the hash workers. I think the proxies that pass data up to the DistributedHash container Chare are colliding, because they are all running on the same Chare/PE, and so the messages are getting delayed.