Initial attempt to implement concurrent hash queries (not working fully!)
This is a complete mess, and it's not working fully yet. When poset nodes and hash nodes are on orthogonal sets of PEs it seems to work fine with little overhead. Also, when the poset/hash PEs are congruent it seems to work consistently, but with a significant (~0.1 sec) penalty. I need to test this to see if it's a fixed setup-time issue, however I doubt it (when using orthogonal sets of PEs, the execution time increases by *less* than a factor of 1 over the number of PEs used).