[Empty commit] Follow up on previous commit: no performance regression
It seems the performance regression was because I was trying only single examples at a time instead of my usual 129-example ACAS test bench. There is some one-time setup cost of the whole program that was being reflected in the resulting times.

However, for certain non-orthogonal sets of PEs, deadlock situations are still occurring; no deadlocks when posetPEs==hashPEs, though