Fix nasty synchronization bug
NOTE TO SELF: even recv() methods allow a coroutine to suspend, apparently. I was setting self.initiatedQueryProc and self.initiatedNodeProc **after** recv() methods in both cases. This meant that the recv() method was allowing listen() to suspend, upon which localListeners were setting that property to True. After the recv() call, I was then *resetting* that property to False, thereby ensuring that an expected message was not sent by localQueryListener/localListener.

Yikes! This was a tough one to debug, but thank goodness I finally decided to use the formatted output I did (with the common PE number appearing in a single column in the output).