Fix use of queries to correctly skip LPs in posetFastLP
Note: the overhead of running queries seems to be considerable, at least when the queries are on the same PE as the successorWorkers. I suspect that an enormous amount of time is wasted with the deferControl calls waiting for the query mutex. Since individual LPs execute very quickly, I suspect they don't get "staggered" at all, and basically we are just watching the successor workers take turns waiting for the query mutex. This is further supported by the increased run time, which looks about as slow as a 2-3 core implementation.

Probably queries should be reserved for the case when the hash workers can reside on their own PEs.

Note, however: for large posets, some significant number of LPs may be saved by this querying procedure, although perhaps not many (on the small "large" ACAS example, the number goes from 9200 -> 8800, for example).