"Improve" concurrency of upper-bound checker (regression: more LPs?)
It seems that this "improvement", which was designed to reduce the number of LPs used, actually results in slightly more being used. This seems to be because it takes an extra "cycle" to distribute the message to the workers that they need to shut down, and one or two more LPs sneak in per verification as a result.

However, I guess this cost is negligible with the consequence of having considerably more readable code. Also, it does seem to improve execution times, although I have only tested on very small problems so far, so this is hard to discern.