Fix computation of search direction
We start with the first variable to "stay", and project its hyperplane normal onto the hyperplane normal of the vector to leave. Then we continue removing components in the direction of all the other "stay" variables. This requires projecting not only onto the hyperplane normal of the leave variable, but all of the components we have already remove by subtracting scaled versions of the other stay variables (think Gram-Schmidt, etc.)