Fix multiple targetHyperplanes argument, and related termination logic
I was thinking that it would be necessary to set targetHyperplanes in different ways, but I think that will prove unnecessary. The reason is that there are other ways to cause early termination if targetHyperplane is redundant (e.g. setting activeConstraints = [], which has length 0). Also, there should never be any need to change the targetHyperplane within a recursive search; that was a mistake. Nevertheless, we will continue to have pivot() set that argument to recursive calls of searchAndFilter().

Also fixed in this commit is the termination condition when there isn't a positive gradient with respect to the target hyperplane, i.e. when it is obviously redundant