More or less the code submitted to HSCC2022
Note that for a number of experiments actual source code of charm4py had to be modified, i.e. the channel buffer size in channel.py needed to be increased. This is due to rate control issues -- i.e. the lack of rate control between successor workers and hash workers.