Attempt to improve performance of previous commit
At least on my small test bench, the previous commit actually *hurt* performance rather than helping it as intended.

This represents the culmination of multiple attempts to improve performance while keeping the bytes/bytearray mechanism of hashing a region. At the end of the day, I think the performance of this methodology is just permanently handicapped relative to *integer* encoding of regions **ON SMALL PROBLEMS WHERE THE NUMBER OF HYEPRPLANES FITS IN A MACHINE-SIZE INTEGER**. I still have to test this on larger problems to see if there is an advantage there.

However, some careful thought has suggested to me that using a purely integer/bytearray encoding of the region is not optimal, even memory-wise. For small numbers of hyperplanes, this methodology has a clear memory advantage. But for large numbers of hyperplanes, **ESPECIALLY WHEN WE LIKELY WON'T ENUMERATE ALL REGIONS** it is unlikely that the number of flipped hyperplanes will come withing 1/64th of the number of hyperplanes itself; thus, keeping integers for the active hyperplanes should even be more memory efficient, not to mention forestalling a pass over integer/bytearray to find which hyperplanes are active.