Re-implement new hashing mechanism; transmit list of *flipped* hyperplanes to distributed hash
This should be both memory efficient (most of the time) and fast, since we no longer do a full pass over a list of length N. The memory efficiency should come from the observation last time that in large problems we will never flip anything close to N/64 hyperplanes.

As it turns out, the performance problem I observed has nothing to do with this: it was some sloppiness in specifying the LP. So that change will be in the next commit. I have to go back and make sure that one of the earlier versions doesn't run faster than this one, after this savings is implemented.