Implement new node-checking mechanism: send nodes back to poset workers for checking
This change is implemented by giving the hash workers access to the poset's successor worker proxy, so that when they have a node to check, they can send it to one of those PEs.

In future, I think the right thing to do is to send it to a *random* successor worker PE to ensure proper load balancing. This will be important because only newly discovered nodes need to be checked, and they should appear unpredictably from successor workers.

Finally, in testing, I noticed the 'False' early termination warnings. So that still needs to be debugged in this edition of the code.

** NB: **
In the course of implementing this mechanism, I discovered that charm4py doesn't handle class inheritance for Chares with ***ANY*** Chares that have custom constructors!!