Allow poset Chare to be reset and recalled on other problems (fixes memory leaks)
Investigations show that Chares are *never* freed once they are created, and this is the cause of the memory leaks experienced with running Fast BATLLNN on a test suite.

The only solution to this problem is re-use a single poset Chare (and dependents) for different problems. This commit does the refactoring necessary to achieve this (although I had a good head start, since I anticipated partial reuse in the context of changing the "bound" associated with Fast BATLLNN verification).

Note: a number of synchronization problems had to be solved as a result. In particular, it became necessary to make sure that all listeners associated with the distributed hash are shut down, so that the same channels can be reused by a subsequent calls for different problems. This became especially problematic with respect to the recently introduced changes which schedule node checking on poset PEs (the asynchronous nature of it was the problem). However, this code was tested on and seems to also function perfectly fine on the original version, which does checks on the hash workers.