Add clearHashTable/resetLevelCount methods (also fix root node check bug)
These methods explicitly clear the hash table and reset the hash table's level count, respectively.

This commit also fixes a bug where the root node was not being checked, because I was not calling getLevelList, the only method that used to clear the hash table. So a hash table that *only* had the root node, was effectively being preserved between different instantiations, preventing the root node from being checked.