Reduce memory usage by copying current level to poset PEs in chunks
In the old implementation, the newly discovered nodes were combined into a single list on the main PE, and then copied to the successor workers from that list. This effectively created a total of four(!!) transient copies of the next level: the one in the table, the one in levelList, the centralized one described above, and the copies distributed to the successor workers. This obviously increases the memory usage of the program by quite a bit. (Note, however, that each of these copies has somewhat different information: e.g. the nodes in the table contain more information than those in the levelList.)

This commit reduces the peak memory usage of the poset in several ways:

1) the current level is tracked using only references to the node instead of a copy of the region's representation (i.e. self.nodeBytes) -- this should save lots of memory by itself, since references are smaller than region representations in large problems;

2) the level lists are copied directly from their hash PE to the destination poset PE, thus eliminating the centralized copy;

3) nodes are copied to their destination PE in chunks, and those nodes are removed from the levelList after each chunk is copied -- this should further free up memory (this is the default setting of clearTable='speed');

4) for extreme situations, setting clearTable='memory' removes nodes from the hash table after a chunk is copied (in addition to the level list), thus ensuring that there are duplicates of at most (# poset PEs * chunk size) nodes -- note that this will be slower, because removing these from the hash table will have some additional cost.