Change option passing to populatePoset
I want to take arbitrary additional named arguments to populatePoset (via **kwargs), and pass them the self.succGroup.setMethod proxy call (again via **kwargs).

However, it seems that charm4py is doing some filtering of the named arguments to the proxy call to populatePost, so **kwargs isn't being passed on to setMethod. This is probably because of ret= and awaitable= named arguments that all remote method executions accept.

Thus, to allow arbitrary named arguments to be automatically passed to setMethod, the best option seemed to be to take a single dictionary with all of the arguments in it. Otherwise, I would have had to split the arguments between those that for populatePoset the method, and those that are to be passed along to setMethod. This distinction would be artificial since all of the arguments affect the behavior/output of the "populate poset" algorithm. Hence my choice here.