Allow hash table to store regions by vertex & list of active flipped hyperplanes
That is, each region is encoded by one of its vertices and the list of active hyperplanes _at that vertex_ which have been flipped to form an associated region. This should considerably reduce the memory usage of the hash table for arrangements with many hyperplanes, since each region is now encoded using O(d) bits; of course this comes at the expense of computation, since each region now requires O(n*d) operations to "decode" into its list of hyperplane-flip bits (more on this below).

In practice, this implementation seems to be considerably slower than the bytearray version of the hash table (hashStore:'bits'), but this is probably reflecting the very low constants of bytearray operations.

This requires an ugly implementation, since it requires "decoding" the stored region via the constraints matrix. That is, essentially finding all of the original hyperplanes for which the vertex lies on the negative side. This has to be done for every _equality_ comparison between nodes, since the same region may have _many_ vertex/active flips encodings!

Ultimately, this meant spoiling the "purity" of the DistributedHash Chare with hyperplane specific information, so that these comparisons could be done within the hash table chares. It also necessitated adding some special paths in several poset chare methods.

On the other hand, it does incorporate some new functionality: now the node comparison function can be user-specified using the fn= optional argument to the new Chare method updateNodeEqualityFn. (This method uses special argument patterns to swithch between the ordinary region encodings -- i.e. bits/lists -- and the new vertex-based region encoding.)