Deprecate findAll and useBounding options
The findAll option currently isn't implemented (and it was intended to
be about returning *all* of the faces of a region). The use bounding
option doesn't make sense in either box-corner implementation or LP
implementation. In the former case, there is exponential growth in the
dimension (although so is the number of regions), and this isn't memory
practical or computation practical for anything but small dimensions. in
the latter case, we are replacing one LP to decide feasibility with
**TWO** LPs (at least!) to decide intersection with the bounding box!!