Fix interpretation of translateRegion and regionInteriorPoint
This was a bone-headed error. The method translateRegion was meant to convert the poset region -- which is based on flippings so that self.pt is on the pos. side of all hyperplanes -- into flippings of the original hyperplanes, as specified.

This is helpful for FastBATLLNN, for example, where we care about the signs of the original hyperplanes, since we care about those as actual linear functions not just regions specifiers (i.e. the region doesn't change by negation, but the linear function does).

The problem is that this method was being used internally to regionInteriorPoint, which necessitated a double flipping to get things right: see subsequent nonsense in the next commit to FastBATLLNN.

Now, however, getRegionConstraints correctly expects as input the region specification from the poset, i.e. relative to the hyperplanes flipped relative to self.pt. A further option allN=True was added, so that in flippedConstraintsReducedMin objects, we can correctly get back the desired information for the redundant constraints as well, if desired. However, for flippedConstraintsReducedMin objects, allN is always set to False, because why find an interior point using redundant constraints?