Correct serious concurrency bug in DistributedHash with poset checking
When scheduling the check method of the localVars group, it is difficult
to ensure that that method has finished executing before terminating a
level. This function was intended to be accomplished with the
`awaitPending` method. However, with multiple `hashAndSend` calls in
flight, it may be possible for `awaitPending` to see a total `schedCount`
of zero, even when there are check methods that haven't finished executing.
As a result, the subsequent termination signal of -2 may be received on
the local listener before all messages pertaining to hash nodes are even
sent. Obviously this is a problem. (FWIW, it definitely seems to be
specific to the case when multiple `hashAndSend` calls are in flight --
I don't think it was possible in the original code otherwise. However,
this is clearly a feature that I expect to use, so it's work fixing...)

The solution to this problem is to have `hashAndSend` increment a local
counter at the beginning of each call. On the DistributedHash, then,
there is a corresponding counter tracking each node that is actually
processed in the hash (including, say, for node updates). Thus, a
revised `awaitPending` method may now compare those two counters to know
whether all pending `hashAndSend` calls have completed successfully