Finally fix concurrency issues with poset checking
Basically what was happening was that in *between* the two counter
retrievals in `awaitPending` other stuff was scheduled, hence leading
`awaitPending` to prematurely conclude that checking was done.

The fix is to use a single counter to track the number hashes vs.
checks/updates that have been done. This entails using the new
`decHashedNodeCount` methods to perform this decrementing on successor
group Chare where the original hash was scheduled.