Provide poset Chare a method for clearing the table
Also note a comment about unintuitive behavior of poset Chare: if
'clearTable' option is set, then the *entire* table is cleared at the
beginning of each populatePoset call, since the final table won't
contain all of the nodes anyway.

However, if 'clearTable' is not set, then this clearing *doesn't* happen,
necessitating an explicit clearing of the hash table is that is desired.
Hence, the additional method.

I'm not convinced of the logic of this behavior, so it may change in the
future...