Begin correct identification of faces split by inserted hyperplane
The idea here is to use the existing `projConstraints` function to create
the hyperplane arrangement projected onto the inserted hyperplane. Using
`lpMinHRep` on this arrangement allows us to see which full-dimensional
hyperplane faces of this region in the original arrangement are split by
the inserted hyperplane. Note that we use the variable `validFlips` to
bookkeep which faces are flippable from the base region containing
`rebasePt`, as well as reading the full-dimensional face information from
the existing arrangement enumeration. This reduces the number of LP
calls we need to use to make this determination.

Note also: this technique takes good advantage of the `vectorSet`
abstraction that now underlies `flipConstraints` objects. When multiple
hyperplanes in the original arrangement have a full-dimensional
intersection in the *projected* arrangement, the `vectorSet` handling of
"equivalent" hyperplanes, combined with its `expandDuplicates` method,
allows us to easily identify and handle the situation. (The redundancies
are automatically removed to make the min H-Rep call efficient, and then
the multiple intersections are handled via `expandDuplicates` afterward.)