Correctly seed initial post-insertion region
`populatePoset` method pulls its value of N (number of hyperplanes) from
its flipped constraints object, which has the post-insertion number of
hyperplanes. Hence, we need to replace the node corresponding to
the projected sub-arrangement's root (as we have chosen it), so that the
default code in `populatePoset` correctly uses this region.

Recall: region comparison now depends on matching `N`, the number of
encoded hyperplanes.