Fix bug in adjFaceCreate: populate node adj dicts only when empty
If no hyperplanes have yet been inserted into a poset table, then the
adj dictionaries of each node should be empty. However, the
insertHyperplane call requires these to be populated with information
about the number of hyperplanes used to encode adjacent regions. Hence,
the adjFaceCreate call.

However, this adjFaceCreate call should be guarded to *only* be called
on the first hyperplane insertion, and it definitely should not update
adj dictionaries that already exist.