Reorder node creation (post insertion) and node deletion
Previous implementation would have generated bugs in dimensions d >= 3.
This is because the same successor region could share faces with
multiple regions at the current level.

In particular, the previous implementation deleted the old,
pre-insertion successor region the *first* time it was encountered. This
would create a bug because another region at the same level would try to
update the adjacency information in the old, pre-insertion region after
it was deleted, thus effectively recreating it.

This bug is not apparent in my current test arrangement, since it is in
d = 2, and hence the projected arrangement is in d = 1 -- i.e. a line.
Thus, successor regions cannot share faces with multiple regions from
the previous level.