Initial 'working' version for non-degenerate arrangements
This should work for non-degenerate arrangements in any dimension, but
it does not yet correctly handle the case where the inserted hyperplane
cuts through the intersection of two or more existing hyperplanes. (That
will require a bit more work.)

I was concerned that this implementation would suffer from concurrency
bugs, specifically with regard to nondeterministic ordering of when old,
N-1 encoded nodes at the current level, and adj updates to all face
nodes. However, I believe this is not an issue, since the defining
characteristic of a 'level' in the poset is that each node contains
exactly the same number of 'flipped' hyperplanes from the root node
(which is specified no flipped hyperplanes); note that we proceed in
the 'rebased' canonicalization of regions. Thus flipping exactly one
face (in either direction) on nodes at the current level necessarily
moves out *out* of that level. Hence there is no possibility of trying
to update a node (in the current level) after it has been deleted by its
associated processNodeSuccessor.