Allow populatePoset to respond to 'maxLevels' option
This option restricts (or extends) the number of levels that
populatePoset is willing to iterate over.

This addition was necessary because of the way that the poset is seeded
in insertHyperplane mode, which results in the need to iterate over one
additional level than the original default setting of N+1.