Fix bug in node updateForInsert method
Basically the automatic update of the `faces` property was implemented
incorrectly for insert mode. This bug was emptying said property for
inserted post-split nodes nodes in insert mode. The effect was thus that
newly inserted nodes were not tracking all of their "incoming" split
faces, which lead to face information corruption.