Correct behavior for degenerate arrangements
Note that the "incoming" faces for created nodes now consist *only* of
those faces that are split into two full-dimensional faces. These do not
need to be checked to find whether they are faces of the positive
and negative split regions, since they are guaranteed to be faces of
both.

All other full-dimensional faces necessarily belong to the positive or
the negative split regions but not both.