Begin code to "canonicalize" a poset table created from inserting hypers
Recall: the hyperplane insertion code does not touch all of the existing
regions in the arrangement. So after (multiple) insertion(s) nodes in
the table will be encoded using various numbers of hyperplanes.

The purpose of this code is re-encode all of the nodes using the final
number of hyperplanes (hence "canonicalize").