Don't iterate over arrays of channels (PEs) in Hash listeners
Instead track the number of pending messages and "done" statuses
(negative status codes) using integers that are incremented/decremented
as necessary.

This should dramatically improve the performance of particularly the
listeners on clusters with a large number of cores, since those
channel/status arrays are ~ to the number of PEs used. Thus, iterating
over those arrays means that the listeners are doing work that scales
linearly with the number of cores! This is not good for the clusters
that I'm planning to deploy with ~700 cores :-)