Finalize code to "canonicalize" a poset created by inserting hyperplanes
In particular, the `canonicalizeTable` method ensures that all entries
in the poset table are encoded with respect to all of the
`self.flippedConstraints.N` hyperplanes.

This method also takes a `rebasePt` optional argument, which ensures
that the re-encoded regions are computed with respect to the base point
`rebasePt`. This is useful for comparison purposes, since creating a
`flippedConstraints` object by inserting hyperplanes may produce a
different initial point (i.e. `self.flippedConstraints.pt`) than the one
produced by creating a constraints object from all of the hyperplanes
together.