First version of `filterParallel` methods
This method takes a query hyperplane (in CDD format), and filters the
non-redundant hyperplanes in the constraint object, by removing the
hyperplanes that are identical/parallel to the query.

This method returns `None` when there are no identical/parallel
hyperplanes in the constraints object and a three-tuple otherwise. The
three-tuple contains: first, a list of identical hyperplanes removed
(should have length <=1 always); second, a list of the parallel
hyperplanes removed; and third a list of indexes (into the non-redundant
constraints) of the hyperplanes that are NOT identical/parallel to the
query.

NOTE: these return values are all with respect to the "reduced" or
non-redundant hyperplane constraints.