Add maskHyperplane method to flipConstraints object
This method effectively removes additional hyperplanes from the
`constraints` list relative to `allConstraints`. It does this by
manipulating `nonRedundantHyperplanes` in much the same way the
`flipConstraintsReducedMin` object does for hyperplanes that do not
intersect the fixed constraint polytope.

Note: the argument is a list of hyperplanes to *keep*. All hyperplanes
not already ignored (see e.g. `flipConstraintsReducedMin`) and *not* in
the input list will be masked.